move getcouncildistrcits
