+++ moped-editor/src/views/projects/projectView/ProjectSummary/ProjectSummaryCouncilDistricts.js
@@ -1,20 +1,7 @@
 import React from "react";
 import { Box, Grid, Typography } from "@mui/material";
+import getCouncilDistricts from "src/utils/getCouncilDistricts"
 
-  // reduce the array of geography objects into an array of city council districts
-  const getDistricts = (data) => {
-    const initialValue = [];
-    const districts = data.reduce(
-      (acc, component) => [...acc, component["council_districts"]],
-      initialValue
-    );
-
-    // flatten the array of arrays and remove empty districts
-    const districtsArray = districts.flat().filter(d=>d)
-
-    // sort in ascending order and use Set to only return unique districts
-    return [...new Set(districtsArray.sort((a, b) => a - b))];
-  };
 
 /**
  * ProjectSummaryCouncilDistricts Component
@@ -32,7 +19,7 @@ const ProjectSummaryCouncilDistricts = ({ projectGeography, classes }) => {
       </Typography>
       <Box className={classes.fieldBox}>
         <Typography className={classes.fieldLabelTextNoHover}>
-          {getDistricts(projectGeography).join(", ")}
+          {getCouncilDistricts(projectGeography).join(", ")}
         </Typography>
       </Box>
     </Grid>
